Subclass based on requested shape
